Sonar (#370 chunk H): split cms/services.py into a package
`cms.services` was a 3,584-line module; Sonar python:S104 caps it at
500. Split into a package with focused submodules and a thin
re-exporting facade so existing `from cms.services import X` call
sites and `unittest.mock.patch("cms.services.X")` targets keep
working:

- `cms/services/_common.py` (~245) — shared validators
  (`_validate_caller_user`, `_validate_nonneg_int_id`,
  `_validate_listing_user`, `_validate_nonempty_str`,
  `_validate_positive_int`), the agent-projection shape table /
  `_agent_projection_dict`, plus `_flatten_range_spec_instances` /
  `_instance_contexts_from_range_spec` / `_resolve_runtime_ips` (the
  PEP 695-generic instance-context builder used by every range
  projection).
- `cms/services/_agents.py` (~310) — create/delete/list/get_agent +
  get_allowed_extensions; late-bound `assets_create_agent` /
  `assets_delete_agent` / `get_agent` lookups so tests that patch
  `cms.services.X` keep working.
- `cms/services/_credentials.py` (~455) — create/delete/list/get
  credential.
- `cms/services/_range_queries.py` (~335) — list_ranges, get_range,
  get_active_range, get_range_by_request_id.
- `cms/services/_range_create.py` (~320) — create_range + the
  validate/load/check/lookup/dispatch/persist/audit/context-build
  helpers; late-bound `engine_create_range` / `audit_log` /
  `get_active_range` / `get_agent` calls.
- `cms/services/_range_destroy.py` (~455) — destroy_range,
  cancel_range, destroy_range_by_request_id,
  cancel_range_by_request_id; late-bound engine + audit_log calls.
- `cms/services/_range_pause.py` (~245) — pause_range,
  pause_range_by_request_id.
- `cms/services/_range_resume.py` (~245) — resume_range,
  resume_range_by_request_id.
- `cms/services/_uploads.py` (~420) — initiate_upload (+ the
  validate-inputs / inner / quota-and-token helpers), complete_upload
  (+ `_verify_upload_token_or_raise`, `_verify_upload_object_or_raise`,
  `_inspect_upload_header_or_raise`), cancel_upload, get_storage_used.
- `cms/services/_scenarios.py` (~135) — list_scenarios, get_scenario,
  validate_scenario_requirements.
- `cms/services/_ngfws.py` (~475) — list_ngfws, get_ngfw, create_ngfw
  + `_validate_ngfw_*`, `_resolve_ngfw_*`,
  `_reject_existing_active_ngfw`, `_provision_ngfw_request_records`,
  `_hydrate_and_dispatch_ngfw`, `_app_to_ngfw_context`,
  `_validate_app_id`, destroy_ngfw.
- `cms/services/_queries.py` (~70) — system-level read-only range
  queries used by CTF (get_range_status_by_id, get_range_spec_by_id,
  find_range_instance_id_by_request, get_range_target_instances).
- `cms/services/__init__.py` (~160) — re-exports the public API plus
  the historically-patchable names (`assets_create_agent`,
  `assets_delete_agent`, `audit_log`, `engine_create_range`,
  `engine_destroy_range_by_request`, `engine_cancel_range_by_request`,
  `engine_pause_range`, `engine_resume_range`,
  `engine_get_instance_ips_by_uuid`, `RangeInstance`) and the cross-
  layer re-exports the layer-imports gate depends on
  (`ScriptUploadError`, `complete_script_upload`, `delete_script`,
  `initiate_script_upload`, `list_scripts`, `range_status_changed`).

All 3,212 platform tests pass post-split; no behaviour change. Layer-
imports check (mission_control / ctf -> cms.services only) still
passes.
