+++ shifter/shifter_platform/cms/services/_range_pause.py
@@ -1,0 +1,242 @@
+"""Range pause entrypoints (by range_id and by request_id)."""
+
+from __future__ import annotations
+
+import logging
+from typing import TYPE_CHECKING, Any
+
+from cms.exceptions import CMSError
+from cms.models import RangeInstance
+from risk_register.models import AuditLog
+from shared.constants import USER_CANNOT_BE_NONE
+from shared.enums import ResourceStatus
+
+from ._common import _validate_caller_user
+
+if TYPE_CHECKING:
+    from django.contrib.auth.models import User
+
+logger = logging.getLogger(__name__)
+
+
+def _engine_pause_range_call(request_id: Any) -> Any:
+    """Late-bound call so test patches of cms.services.engine_pause_range apply."""
+    from cms import services as _cs
+
+    return _cs.engine_pause_range(request_id)
+
+
+def _audit_log_call(**kwargs: Any) -> None:
+    """Late-bound call to ``cms.services.audit_log`` so test patches apply."""
+    from cms import services as _cs
+
+    _cs.audit_log(**kwargs)
+
+
+def pause_range(user: User, range_id: int) -> None:
+    """Pause a running range.
+
+    Fetches RangeInstance, verifies ownership, updates CMS status to PAUSING,
+    then delegates to engine.services.pause_range.
+
+    Args:
+        user: User requesting pause
+        range_id: ID of the range to pause
+
+    Returns:
+        None
+
+    Raises:
+        TypeError: If user is None, invalid type, or range_id is invalid type
+        ValueError: If user has no ID (unsaved) or range_id is invalid
+        CMSError: If range not found, not owned by user, or not in pausable state
+    """
+    _validate_caller_user(user, "pause_range")
+
+    if range_id is None:
+        logger.error(
+            "pause_range called with None range_id for user_id=%s",
+            user.id,
+        )
+        raise TypeError("range_id cannot be None")
+
+    if not isinstance(range_id, int):
+        logger.error(
+            "pause_range called with invalid range_id type: %s",
+            type(range_id).__name__,
+        )
+        msg = f"range_id must be an int, got {type(range_id).__name__}"
+        raise TypeError(msg)
+
+    if range_id < 0:
+        logger.error(
+            "pause_range called with negative range_id=%s for user_id=%s",
+            range_id,
+            user.id,
+        )
+        raise ValueError("range_id must be non-negative")
+
+    logger.debug(
+        "pause_range called for user_id=%s, range_id=%s",
+        user.id,
+        range_id,
+    )
+
+    try:
+        instance = RangeInstance.objects.get(range_id=range_id)
+    except RangeInstance.DoesNotExist:
+        logger.warning(
+            "pause_range: range not found for user_id=%s, range_id=%s",
+            user.id,
+            range_id,
+        )
+        raise CMSError(f"Range {range_id} not found") from None
+
+    if instance.user_id != user.id:
+        logger.error(
+            "pause_range: access denied - range_id=%s owned by user_id=%s, requested by user_id=%s",
+            range_id,
+            instance.user_id,
+            user.id,
+        )
+        raise CMSError(f"Range {range_id} not found")
+
+    try:
+        request_id = instance.request.request_id if instance.request else None
+        if request_id is None:
+            logger.error(
+                "pause_range: no request_id for range_id=%s, cannot pause",
+                range_id,
+            )
+            raise CMSError("Range has no associated request")
+
+        instance.status = ResourceStatus.PAUSING.value
+        instance.save(update_fields=["status"])
+
+        if not _engine_pause_range_call(request_id):
+            instance.status = ResourceStatus.READY.value
+            instance.save(update_fields=["status"])
+            logger.warning(
+                "pause_range: engine returned False for range_id=%s",
+                range_id,
+            )
+            raise CMSError("Range cannot be paused in current state")
+
+        _audit_log_call(
+            entity_type=AuditLog.EntityType.RANGE,
+            entity_id=range_id,
+            action=AuditLog.Action.PAUSE,
+            actor_type=AuditLog.ActorType.USER,
+            actor_id=user.id,
+            new_state={"status": ResourceStatus.PAUSING.value},
+            request_id=str(request_id),
+        )
+
+        logger.info(
+            "pause_range completed: range_id=%s user_id=%s",
+            range_id,
+            user.id,
+        )
+    except (TypeError, ValueError, CMSError):
+        raise
+    except Exception:
+        logger.exception(
+            "Error in pause_range: user_id=%s range_id=%s",
+            user.id,
+            range_id,
+        )
+        raise
+
+
+def pause_range_by_request_id(user: User, request_id: str) -> None:
+    """Pause a running range by request_id.
+
+    Fetches RangeInstance by request_id, verifies ownership, then delegates
+    to engine.services.pause_range.
+
+    Args:
+        user: User requesting pause
+        request_id: UUID string of the request
+
+    Returns:
+        None
+
+    Raises:
+        TypeError: If user is None or invalid type
+        CMSError: If range not found, not owned by user, or not in pausable state
+    """
+    if user is None:
+        logger.error("pause_range_by_request_id called with None user")
+        raise TypeError(USER_CANNOT_BE_NONE)
+
+    if not hasattr(user, "id"):
+        logger.error(
+            "pause_range_by_request_id called with invalid user type: %s",
+            type(user).__name__,
+        )
+        msg = f"user must be a User instance, got {type(user).__name__}"
+        raise TypeError(msg)
+
+    if not request_id:
+        logger.error("pause_range_by_request_id called with empty request_id")
+        raise CMSError("request_id is required")
+
+    logger.debug(
+        "pause_range_by_request_id called: user_id=%s request_id=%s",
+        user.id,
+        request_id,
+    )
+
+    instance = RangeInstance.objects.filter(
+        request__request_id=request_id,
+        user_id=user.id,
+    ).first()
+
+    if not instance:
+        logger.warning(
+            "pause_range_by_request_id: not found: request_id=%s user_id=%s",
+            request_id,
+            user.id,
+        )
+        raise CMSError("Range not found")
+
+    if instance.request is None:
+        raise CMSError("Range has no associated request")
+
+    try:
+        instance.status = ResourceStatus.PAUSING.value
+        instance.save(update_fields=["status"])
+
+        if not _engine_pause_range_call(instance.request.request_id):
+            instance.status = ResourceStatus.READY.value
+            instance.save(update_fields=["status"])
+            logger.warning(
+                "pause_range_by_request_id: engine returned False for request_id=%s",
+                request_id,
+            )
+            raise CMSError("Range cannot be paused in current state")
+
+        _audit_log_call(
+            entity_type=AuditLog.EntityType.RANGE,
+            entity_id=instance.range_id or 0,
+            action=AuditLog.Action.PAUSE,
+            actor_type=AuditLog.ActorType.USER,
+            actor_id=user.id,
+            new_state={"status": ResourceStatus.PAUSING.value},
+            request_id=str(instance.request.request_id),
+        )
+
+        logger.info(
+            "pause_range_by_request_id completed: request_id=%s user_id=%s",
+            request_id,
+            user.id,
+        )
+    except (TypeError, ValueError, CMSError):
+        raise
+    except Exception:
+        logger.exception(
+            "Error in pause_range_by_request_id: user_id=%s request_id=%s",
+            user.id,
+            request_id,
+        )
+        raise
